# services/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample Ocserv User management Api server.",
        "title": "Ocserv User management Example Api",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/ocserv/groups": {
            "get": {
                "description": "List of Ocserv groups",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ocserv(Groups)"
                ],
                "summary": "List of Ocserv groups",
                "parameters": [
                    {
                        "minimum": 1,
                        "type": "integer",
                        "description": "Page number, starting from 1",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "maximum": 100,
                        "minimum": 1,
                        "type": "integer",
                        "description": "Number of items per page",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Field to order by",
                        "name": "order",
                        "in": "query"
                    },
                    {
                        "enum": [
                            "ASC",
                            "DESC"
                        ],
                        "type": "string",
                        "description": "Sort order, either ASC or DESC",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Bearer TOKEN",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/ocserv_group.OcservGroupsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Unauthorized"
                        }
                    }
                }
            },
            "post": {
                "description": "Ocserv Group creation",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ocserv(Groups)"
                ],
                "summary": "Ocserv Group creation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer TOKEN",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "ocserv group create data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ocserv_group.CreateOcservGroupData"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.OcservGroup"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Unauthorized"
                        }
                    }
                }
            }
        },
        "/ocserv/groups/defaults": {
            "get": {
                "description": "Ocserv Defaults Group config",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ocserv(Groups)"
                ],
                "summary": "Ocserv Defaults Group config",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer TOKEN",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Unauthorized"
                        }
                    }
                }
            },
            "patch": {
                "description": "Update Ocserv Defaults Group",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ocserv(Groups)"
                ],
                "summary": "Update Ocserv Defaults Group",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer TOKEN",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "ocserv group default data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ocserv_group.UpdateOcservGroupData"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Unauthorized"
                        }
                    }
                }
            }
        },
        "/ocserv/groups/lookup": {
            "get": {
                "description": "List of Ocserv group names",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ocserv(Groups)"
                ],
                "summary": "List of Ocserv group names",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer TOKEN",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Unauthorized"
                        }
                    }
                }
            }
        },
        "/ocserv/groups/{id}": {
            "get": {
                "description": "Ocserv group detail",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ocserv(Groups)"
                ],
                "summary": "Ocserv group detail",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Ocserv Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Bearer TOKEN",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.OcservGroup"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Unauthorized"
                        }
                    }
                }
            },
            "delete": {
                "description": "Ocserv Group delete",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ocserv(Groups)"
                ],
                "summary": "Ocserv Group delete",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer TOKEN",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Ocserv Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Unauthorized"
                        }
                    }
                }
            },
            "patch": {
                "description": "Ocserv Group update",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Ocserv(Groups)"
                ],
                "summary": "Ocserv Group update",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer TOKEN",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Ocserv Group ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ocserv group create data",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/ocserv_group.UpdateOcservGroupData"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/models.OcservGroup"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/request.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/middlewares.Unauthorized"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "middlewares.Unauthorized": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string"
                }
            }
        },
...
        "models.OcservGroup": {
            "type": "object",
            "required": [
                "name",
                "owner"
            ],
            "properties": {
                "config": {
                    "$ref": "#/definitions/models.OcservGroupConfig"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "owner": {
                    "type": "string"
                }
            }
        },
        "models.OcservGroupConfig": {
            "type": "object",
            "properties": {
                "cgroup": {
                    "description": "Linux control group to assign the VPN worker process to. Format: 'controller,subsystem:name'. Example: 'cpuset,cpu:test'",
                    "type": "string"
                },
                "deny-roaming": {
                    "description": "Disconnect client if its IP changes (e.g., due to network switch). Example: true",
                    "type": "boolean"
                },
                "dns": {
                    "description": "Comma-separated list of DNS servers to assign to the client. Example: '8.8.8.8,1.1.1.1'",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "dpd": {
                    "description": "Dead Peer Detection timeout in seconds. Example: 90",
                    "type": "integer"
                },
                "explicit-ipv4": {
                    "description": "Static IPv4 address to assign to client. Example: '192.168.100.10'",
                    "type": "string"
                },
                "idle-timeout": {
                    "description": "Time in seconds before disconnecting idle clients. Example: 600",
                    "type": "integer"
                },
                "ipv4-network": {
                    "description": "The pool of addresses that leases will be given from. Example: '192.168.1.0/24'",
                    "type": "string"
                },
                "iroute": {
                    "description": "Internal route available only via VPN. Format: 'IP/prefix'. Example: '10.0.0.0/8'",
                    "type": "string"
                },
                "keepalive": {
                    "description": "Interval in seconds to send keep-alive pings. Example: 60",
                    "type": "integer"
                },
                "max-same-clients": {
                    "description": "Maximum simultaneous logins per user. Example: 2",
                    "type": "integer"
                },
                "mobile-dpd": {
                    "description": "DPD timeout specifically for mobile clients. Example: 300",
                    "type": "integer"
                },
                "mobile-idle-timeout": {
                    "description": "Idle timeout for mobile clients. Example: 900",
                    "type": "integer"
                },
                "mtu": {
                    "description": "Tunnel interface MTU to avoid fragmentation. Example: 1400",
                    "type": "integer"
                },
                "nbns": {
                    "description": "NetBIOS Name Servers (WINS) for Windows clients. Example: '192.168.1.1'",
                    "type": "string"
                },
                "net-priority": {
                    "description": "Priority for routes; lower is higher priority. Example: 1",
                    "type": "integer"
                },
                "no-route": {
                    "description": "List of networks to exclude from VPN routing. Each entry should be in 'IP/prefix' format. Example: ['192.168.0.0/16', '10.0.0.0/8']",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "no-udp": {
                    "description": "Disables UDP, enforcing TCP-only VPN connection. Example: true",
                    "type": "boolean"
                },
                "restrict-user-to-ports": {
                    "description": "Comma-separated list of allowed (or blocked, if negated) protocols and ports. Supports 'tcp(port)', 'udp(port)', 'icmp()', 'icmpv6()', and negation with '!()'. Example: 'tcp(443), tcp(80), udp(53)', or '!(tcp(22), udp(1194))'",
                    "type": "string"
                },
                "restrict-user-to-routes": {
                    "description": "Allow client access only to defined routes. Example: true",
                    "type": "boolean"
                },
                "route": {
                    "description": "Routes pushed to the client for routing traffic. Example: ['0.0.0.0/0', '10.10.0.0/16']",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "rx-data-per-sec": {
                    "description": "Maximum receive bandwidth in bytes per second. Example: '100000' for 100 KB/s",
                    "type": "integer"
                },
                "session-timeout": {
                    "description": "Max session time in seconds before forced disconnect. Example: 3600",
                    "type": "integer"
                },
                "split-dns": {
                    "description": "List of domains over which the provided DNS servers should be used. Example: ['example.com', 'internal.company.com']",
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "stats-report-time": {
                    "description": "Interval in seconds for stats reporting. Example: 300",
                    "type": "integer"
                },
                "tunnel-all-dns": {
                    "description": "Force all DNS traffic through the VPN tunnel. Example: true",
                    "type": "boolean"
                },
                "tx-data-per-sec": {
                    "description": "Maximum transmit bandwidth in bytes per second. Example: '200000' for 200 KB/s",
                    "type": "integer"
                }
            }
        },
...
        "ocserv_group.CreateOcservGroupData": {
            "type": "object",
            "required": [
                "config",
                "name"
            ],
            "properties": {
                "config": {
                    "$ref": "#/definitions/models.OcservGroupConfig"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "ocserv_group.OcservGroupsResponse": {
            "type": "object",
            "required": [
                "meta"
            ],
            "properties": {
                "meta": {
                    "$ref": "#/definitions/request.Meta"
                },
                "result": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.OcservGroup"
                    }
                }
            }
        },
        "ocserv_group.UpdateOcservGroupData": {
            "type": "object",
            "required": [
                "config"
            ],
            "properties": {
                "config": {
                    "$ref": "#/definitions/models.OcservGroupConfig"
                }
            }
        },
...
        "request.ErrorResponse": {
            "type": "object",
            "required": [
                "error",
                "message"
            ],
            "properties": {
                "error": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "message": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "request.Meta": {
            "type": "object",
            "required": [
                "page",
                "size",
                "total_records"
            ],
            "properties": {
                "page": {
                    "type": "integer"
                },
                "size": {
                    "type": "integer"
                },
                "total_records": {
                    "type": "integer"
                }
            }
        },
...
    }
}